MISC: Some warning fixes, vendor updates, api export changes, render contexts.

+++ LinaEngine/src/IMGUI/IMGUILayer.cpp
@@ -199,7 +199,7 @@ namespace LinaEngine
 	bool IMGUILayer::OnWindowResizeEvent(WindowResizeEvent & e)
 	{
 		ImGuiIO& io = ImGui::GetIO();
-		io.DisplaySize = ImVec2(e.GetWidth(), e.GetHeight());
+		io.DisplaySize = ImVec2(static_cast<float>(e.GetWidth()), static_cast<float>(e.GetHeight()));
 		io.DisplayFramebufferScale = ImVec2(1.0f, 1.0f);
 		//glViewport(0, 0, e.GetWidth(), e.GetHeight());
